+++ packages/transaction-controller/src/utils/first-time-interaction.ts
@@ -9,7 +9,7 @@ import {
   type GetAccountAddressRelationshipRequest,
 } from '../api/accounts-api';
 import { projectLogger as log } from '../logger';
-import type { TransactionMeta } from '../types';
+import { TransactionType, type TransactionMeta } from '../types';
 
 type UpdateFirstTimeInteractionRequest = {
   existingTransactions: TransactionMeta[];
@@ -57,10 +57,17 @@ export async function updateFirstTimeInteraction({
     chainId,
     id: transactionId,
     txParams: { data, from, to },
+    type,
   } = transactionMeta;
 
   let recipient;
-  if (data) {
+  if (
+    data &&
+    [
+      TransactionType.tokenMethodTransfer,
+      TransactionType.tokenMethodTransferFrom,
+    ].includes(type as TransactionType)
+  ) {
     const parsedData = decodeTransactionData(data) as TransactionDescription;
     // _to is for ERC20, ERC721 and USDC
     // to is for ERC1155