+++ packages/keyring-controller/src/KeyringController.ts
@@ -219,6 +219,11 @@ export type KeyringControllerUnlockEvent = {
   payload: [];
 };
 
+export type KeyringControllerNewVaultEvent = {
+  type: `${typeof name}:newVault`;
+  payload: [];
+};
+
 export type KeyringControllerActions =
   | KeyringControllerGetStateAction
   | KeyringControllerSignMessageAction
@@ -244,7 +249,8 @@ export type KeyringControllerEvents =
   | KeyringControllerStateChangeEvent
   | KeyringControllerLockEvent
   | KeyringControllerUnlockEvent
-  | KeyringControllerAccountRemovedEvent;
+  | KeyringControllerAccountRemovedEvent
+  | KeyringControllerNewVaultEvent;
 
 export type KeyringControllerMessenger = Messenger<
   typeof name,
@@ -1912,6 +1918,7 @@ export class KeyringController<
 
     await this.#clearKeyrings();
     await this.#createKeyringWithFirstAccount(keyring.type, keyring.opts);
+    this.messenger.publish(`${name}:newVault`);
     this.#setUnlocked();
   }