+++ packages/user-profile-controller/src/UserProfileController.test.ts
@@ -11,6 +11,7 @@ import type { CaipAccountId } from '@metamask/utils';
 
 import {
   UserProfileController,
+  getDefaultUserProfileControllerState,
   type UserProfileControllerMessenger,
 } from './UserProfileController';
 import type { UserProfileUpdateRequest } from './UserProfileService';
@@ -167,6 +168,28 @@ describe('UserProfileController', () => {
 
           expect(pollSpy).toHaveBeenCalledTimes(1);
         });
+      });
+    });
+
+    describe('when KeyringController:newVault is published', () => {
+      it('resets the state to initial state', async () => {
+        await withController(
+          {
+            options: {
+              state: {
+                firstSyncCompleted: true,
+                syncQueue: { someId: ['0xSomeAccount'] },
+              },
+            },
+          },
+          async ({ controller, rootMessenger }) => {
+            rootMessenger.publish('KeyringController:newVault');
+
+            expect(controller.state).toStrictEqual(
+              getDefaultUserProfileControllerState(),
+            );
+          },
+        );
       });
     });
 
@@ -420,6 +443,7 @@ function getMessenger(
     events: [
       'KeyringController:unlock',
       'KeyringController:lock',
+      'KeyringController:newVault',
       'AccountsController:accountAdded',
     ],
   });