+++ packages/user-profile-controller/src/UserProfileController.ts
@@ -9,6 +9,7 @@ import type {
 } from '@metamask/base-controller';
 import type {
   KeyringControllerLockEvent,
+  KeyringControllerNewVaultEvent,
   KeyringControllerUnlockEvent,
 } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
@@ -120,6 +121,7 @@ export type UserProfileControllerEvents = UserProfileControllerStateChangeEvent;
 type AllowedEvents =
   | KeyringControllerUnlockEvent
   | KeyringControllerLockEvent
+  | KeyringControllerNewVaultEvent
   | AccountsControllerAccountAddedEvent;
 
 /**
@@ -237,6 +239,11 @@ export class UserProfileController extends StaticIntervalPollingController()<
       this.stopAllPolling();
     });
 
+    this.messenger.subscribe(
+      'KeyringController:newVault',
+      this.#resetState.bind(this),
+    );
+
     this.messenger.subscribe('AccountsController:accountAdded', (account) => {
       this.#queueAccount(account).catch(console.error);
     });
@@ -289,6 +296,17 @@ export class UserProfileController extends StaticIntervalPollingController()<
           address: accountToCaipAccountId(account),
         });
       });
+    });
+  }
+
+  /**
+   * Resets the controller state to its initial values.
+   * All queued accounts are cleared and the first sync flag is reset.
+   */
+  #resetState() {
+    this.update((state) => {
+      state.firstSyncCompleted = false;
+      state.syncQueue = {};
     });
   }
 }