+++ packages/network-controller/CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - This event is now called `NetworkController:rpcEndpointRetried`.
   - The event payload has been changed as well: `failoverEndpointUrl` has been renamed to `endpointUrl`, and `endpointUrl` has been renamed to `primaryEndpointUrl`. In addition, `networkClientId` and `attempt` have been added to the payload.
 - **BREAKING:** Update `AbstractRpcService`/`RpcServiceRequestable` to remove `{ isolated: true }` from the `onBreak` event data type ([#7166](https://github.com/MetaMask/core/pull/7166))
-  - This represented the error produced when `.isolate` is called on a Cockatiel circuit breaker policy, which we never do.
+  - This represented the error produced when `isolate` is called on a Cockatiel circuit breaker policy, which we don't do (at least, not in the way it is designed to be called)
 - Move peer dependencies for controller and service packages to direct dependencies ([#7209](https://github.com/MetaMask/core/pull/7209))
   - The dependencies moved are:
     - `@metamask/error-reporting-service` (^3.0.0)